moved layout folder
moved layout folder &
updated imports for layout.js
