+++ components/layouts/theme-index-layout.js
