+++ pages/circulaire-windturbines/welke-overheid-heeft.js
@@ -1,4 +1,4 @@
-import Layout from '../../layouts/layout';
+import Layout from '../../components/layouts/layout';
 import WelkeLayout from '../../layouts/welke-layout';
 import WindmillImage from '../../public/Windmill_Measures_Web.png';
 import WindmillImageMob from '../../public/Windmill_Measures_Mob.png';
