+++ pages/privacy-policy/index.js
@@ -1,4 +1,4 @@
-import Layout from '../../layouts/layout';
+import Layout from '../../components/layouts/layout';
 
 export default function Privacy() {
   return (
