+++ pages/wat-is-circulaw/index.js
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import Layout from '../../layouts/layout';
+import Layout from '../../components/layouts/layout';
 import OverNav from '../../components/over-nav';
 
 export default function WatIsCircuLaw() {
